pyproject.toml
[tool.poetry]
name = "glotov"
version = "0.1.0"
description = ""
authors = ["denis sergeevich pekshev <pekshev.ds@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2.23"
djangorestframework = "3.14.0"
mysqlclient = "2.2.0"
slugify = "0.0.1"
pillow = "10.1.0"
django-cors-headers = "4.3.1"
django-filter = "23.5"
pydantic = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
